AppCenter周りの処理をAppCenterServicesに移動
呼び出しコードは別の場所に書く予定のため、とりあえず元のところには書かないでおく。
